Flatten zipline-database and zipline-loader
We don't want a public module called zipline-database because
we're not a database library; that's an implementation detail.
