+++ settings.gradle.kts
@@ -8,8 +8,6 @@ include(":zipline-kotlin-plugin")
 include(":zipline-kotlin-plugin-hosted")
 include(":zipline-kotlin-plugin:tests")
 include(":zipline-bytecode")
-include(":zipline-database")
-include(":zipline-database-testing")
 include(":zipline-loader")
 include(":zipline-loader-testing")
 include(":zipline-profiler")
