--- zipline-database/build.gradle.kts
@@ -1,140 +1,0 @@
-import com.vanniktech.maven.publish.JavadocJar
-import com.vanniktech.maven.publish.KotlinMultiplatform
-import com.vanniktech.maven.publish.MavenPublishBaseExtension
-import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget
-
-plugins {
-  kotlin("multiplatform")
-  id("com.android.library")
-  kotlin("plugin.serialization")
-  id("org.jetbrains.dokka")
-  id("com.vanniktech.maven.publish.base")
-}
-
-kotlin {
-  jvm()
-  android {
-    publishAllLibraryVariants()
-  }
-  if (false) {
-    linuxX64()
-  }
-  macosX64()
-  macosArm64()
-  iosArm64()
-  iosX64()
-  iosSimulatorArm64()
-  tvosArm64()
-  tvosSimulatorArm64()
-  tvosX64()
-
-  sourceSets {
-    val commonMain by getting {
-      dependencies {
-        api(libs.okio.core)
-        api(libs.sqldelight.runtime)
-      }
-    }
-    val engineMain by creating {
-      dependsOn(commonMain)
-    }
-    val jniMain by creating {
-      dependsOn(engineMain)
-      dependencies {
-      }
-    }
-    val jvmMain by getting {
-      dependsOn(jniMain)
-      dependencies {
-        implementation(libs.sqldelight.driver.sqlite)
-        implementation(libs.sqlite.jdbc)
-      }
-    }
-    val androidMain by getting {
-      dependsOn(jniMain)
-      dependencies {
-        implementation(libs.sqldelight.driver.android)
-      }
-    }
-
-    val commonTest by getting {
-      dependencies {
-        implementation(kotlin("test"))
-        implementation(projects.ziplineDatabaseTesting)
-      }
-    }
-    val engineTest by creating {
-      dependsOn(commonTest)
-      dependencies {
-      }
-    }
-    val jniTest by creating {
-      dependsOn(engineTest)
-      dependencies {
-        implementation(libs.junit)
-        implementation(libs.okio.fakeFileSystem)
-        implementation(libs.sqldelight.driver.sqlite)
-        implementation(libs.sqlite.jdbc)
-        implementation(libs.turbine)
-      }
-    }
-    val jvmTest by getting {
-      dependsOn(jniTest)
-      dependencies {
-      }
-    }
-    val nativeMain by creating {
-      dependsOn(engineMain)
-      dependencies {
-        implementation(libs.sqldelight.driver.native)
-        implementation(libs.crashkios)
-      }
-    }
-    val nativeTest by creating {
-      dependsOn(engineTest)
-      dependencies {
-        implementation(libs.sqldelight.driver.native)
-        implementation(libs.crashkios)
-      }
-    }
-    targets.withType<KotlinNativeTarget> {
-      val main by compilations.getting {
-        defaultSourceSet.dependsOn(nativeMain)
-      }
-      val test by compilations.getting {
-        defaultSourceSet.dependsOn(nativeTest)
-      }
-    }
-  }
-}
-
-
-android {
-  compileSdkVersion(libs.versions.compileSdk.get().toInt())
-
-  defaultConfig {
-    minSdkVersion(18)
-    multiDexEnabled = true
-    testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-  }
-
-  sourceSets {
-    getByName("main") {
-      manifest.srcFile("src/androidMain/AndroidManifest.xml")
-    }
-    getByName("androidTest") {
-      java.srcDirs("src/jniTest/kotlin/")
-    }
-  }
-}
-
-dependencies {
-  androidTestImplementation(libs.androidx.test.runner)
-  androidTestImplementation(libs.junit)
-  androidTestImplementation(libs.kotlinx.coroutines.test)
-  androidTestImplementation(libs.okio.fakeFileSystem)
-}
-
-configure<MavenPublishBaseExtension> {
-  configure(KotlinMultiplatform(javadocJar = JavadocJar.Dokka("dokkaGfm")))
-}
