--- zipline-database/src/jvmMain/kotlin/app/cash/zipline/database/databaseJvm.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright (C) 2022 Block, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package app.cash.zipline.database
-
-import com.squareup.sqldelight.db.SqlDriver
-import com.squareup.sqldelight.sqlite.driver.JdbcSqliteDriver
-
-/** JVM implementation is used for testing and thus is In-Memory. */
-actual class DriverFactory(
-  private val schema: SqlDriver.Schema
-) {
-  actual fun createDriver(): SqlDriver {
-    val driver: SqlDriver = JdbcSqliteDriver(JdbcSqliteDriver.IN_MEMORY)
-    schema.create(driver)
-    return driver
-  }
-}
-
-actual fun isSqlException(e: Exception) = e is java.sql.SQLException
