+++ zipline-loader/src/jvmTest/kotlin/app/cash/zipline/loader/DownloadOnlyFetcherReceiverTest.kt
@@ -16,7 +16,7 @@
 package app.cash.zipline.loader
 
 import app.cash.zipline.QuickJs
-import app.cash.zipline.database.DriverFactory
+import app.cash.zipline.loader.internal.database.DriverFactory
 import app.cash.zipline.loader.testing.LoaderTestFixtures
 import app.cash.zipline.loader.testing.LoaderTestFixtures.Companion.alphaUrl
 import app.cash.zipline.loader.testing.LoaderTestFixtures.Companion.bravoUrl
