+++ zipline-loader/src/nativeMain/kotlin/app/cash/zipline/loader/internal/database/databaseNative.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright (C) 2022 Block, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package app.cash.zipline.loader.internal.database
+
+import com.squareup.sqldelight.db.SqlDriver
+import com.squareup.sqldelight.drivers.native.NativeSqliteDriver
+
+actual class DriverFactory(
+  private val schema: SqlDriver.Schema,
+  private val dbName: String,
+) {
+  init {
+    validateDbName(dbName)
+  }
+
+  actual fun createDriver(): SqlDriver {
+    return NativeSqliteDriver(schema, dbName)
+  }
+}
+
+// TODO find the native exception class
+actual fun isSqlException(e: Exception): Boolean = TODO()
